Speech to Text and Text to Speech
I added google Text to Speech so the program will tell you what you said. I will have to fix this so that it does not run on a while loop, and only runs once. To use this program, you have to download google's speech to text, text to speech and pygame in Python3.